feature(simple-api): replaced old code with new API
